Cleanup D2::Plugin->import
Create the registry Core::Plugin object, add the plugin mini DSL to the
caller, optionally import the main DSL keywords and register the plugin
with the calling app. Simple! :P

Core::App uses an arrayref to record registered plugins; its possible
that a plugin could be added mroe than once. Need to look at changing
that to a hashref if this is an issue.